improvement(sourceZipWriter): detach implementation and reduce code smell
    - separate source zip writer implementation from workflowstep.
    - reduce complexity in createZip functions by using stream instead
      of several if statements and loop.
